requirements.txt
pandas
streamlit
snowflake-connector-python
python-dotenv